+++ src/Domain/Metric/MethodBased/CC.php
@@ -4,17 +4,23 @@ declare(strict_types=1);
 
 namespace PHPQuality\Domain\Metric\MethodBased;
 
+/**
+ * @psalm-external-mutation-free
+ */
 final class CC extends MethodBasedMetric
 {
     /** @var int */
     private $complexity = 0;
 
+    /**
+     * Should be called whenever a statement that increases the complexity is found
+     */
     public function increaseComplexity(): void
     {
         $this->complexity++;
     }
 
-    public function calculate(): int
+    public function calculate(): float
     {
         return $this->complexity;
     }
